Side bar is now in it's own .java class
