+++ app/src/main/java/com/example/quintessentiel/ConnectionActivity.java
@@ -1,29 +1,21 @@
 package com.example.quintessentiel;
 
-import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.design.widget.NavigationView;
-import android.support.v4.view.GravityCompat;
-import android.support.v4.widget.DrawerLayout;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
-import android.view.MenuItem;
 import android.view.View;
-import android.view.inputmethod.InputMethodManager;
 import android.widget.EditText;
 
 import com.example.quintessentiel.User.CtrlUser;
 
-public class ConnectionActivity extends AppCompatActivity {
+public class ConnectionActivity extends BaseActivity {
 
     private CtrlUser ctrlUser;
 
     @Override
-    public void onCreate(Bundle savedInstanceState) {
+    protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.connection);
+        super.onCreateDrawer();
 
         this.ctrlUser = new CtrlUser(this);
 
@@ -77,37 +69,7 @@ public class ConnectionActivity extends AppCompatActivity {
             }
         });
 
-        //Side bar
-        Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
-        setSupportActionBar(toolbar);
 
-        findViewById(R.id.drawer_button).setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                // open right drawer
-                DrawerLayout drawer = (DrawerLayout) findViewById(R.id.drawer_layout);
-                drawer.openDrawer(GravityCompat.END);
-
-
-                View view = getCurrentFocus();
-
-                InputMethodManager imm = (InputMethodManager)getSystemService(Context.INPUT_METHOD_SERVICE);
-                imm.hideSoftInputFromWindow(view.getWindowToken(), 0);
-                view.clearFocus(); //The parent of this elements needs to be focusable
-            }
-        });
-
-
-
-        NavigationView navigationView = (NavigationView) findViewById(R.id.nav_view);
-        navigationView.bringToFront();
-        navigationView.setNavigationItemSelectedListener(new NavigationView.OnNavigationItemSelectedListener() {
-            @Override
-            public boolean onNavigationItemSelected(@NonNull MenuItem menuItem) {
-                System.out.println(menuItem);
-                return false;
-            }
-        });
     }
 
     /**
